requirements.txt
Flask==2.3.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
itsdangerous==2.2.0
click==8.3.0
blinker==1.9.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.44
Flask-Migrate==4.1.0
alembic==1.16.5
greenlet==3.2.4
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
bcrypt==5.0.0
Flask-WTF==1.2.2
WTForms==3.2.1
email-validator==2.3.0
Flask-Mail==0.10.0
dnspython==2.8.0
Flask-CKEditor==1.0.0
Pillow==11.3.0
python-dotenv==1.1.1
bleach==6.2.0
Mako==1.3.10
typing_extensions==4.15.0
idna==3.10
